+++ src/Domain/Repository/UserRepositoryInterface.php
@@ -95,6 +95,13 @@ interface UserRepositoryInterface
     public function findWithEnabledSheetByEvent(Event $event): array;
 
     /**
+     * @param Event $event
+     *
+     * @return User[]
+     */
+    public function findOwnersWithEnabledSheetByEvent(Event $event): array;
+
+    /**
      * @param Event  $event
      * @param string $locale
      *