+++ CHANGELOG.md
@@ -5,5 +5,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.0] - 2019-04-17
 ### Added
 - VIMEET-1944: Unallow request meeting placement when linked sheets meeting already placed. 
+
+## [2.0.9] - 2019-04-17
+### Changed
+- VIMEET-1913 - No participants selection on meeting request when meetings are assigned to all participants
+
+## [2.0.8] - 2019-04-17
+### Fixed
+- MV-167 - Fix invoice generator
+
+## [2.0.7] - 2019-04-15
+### Added
+- VIMEET-1937 : Add Matomo tracking code in event extra parameter. Add privacy policy page with Matomo tracking opt-out.
+
+## [2.0.6] - 2019-04-15
+### Fixed
+- MV-168 - Import LENI users ids with owners of sheets